refMetList: remove FXML from controller I
